ensure images are pulled, bump celestia node and ictest

+++ interchaintest/publish_test.go
@@ -47,8 +47,8 @@ func TestPublish(t *testing.T) {
 		AddChain(rollchainChain).
 		AddChain(celestiaChain, ibc.WalletAmount{
 			Address: "celestia1dr3gwf5kulm4e4k0pctwzn0htw6wrvevdgjdlf",
-			Amount: math.NewInt(100_000_000_000), // 100,000 tia
-			Denom: celestiaChainSpec.Denom,
+			Amount:  math.NewInt(100_000_000_000), // 100,000 tia
+			Denom:   celestiaChainSpec.Denom,
 		})
 
 	ctx := context.Background()
@@ -65,7 +65,7 @@ func TestPublish(t *testing.T) {
 		_ = ic.Close()
 	})
 
-	for i:=0; i < nv; i++ {
+	for i := 0; i < nv; i++ {
 		// celestia1dr3gwf5kulm4e4k0pctwzn0htw6wrvevdgjdlf
 		stdout, stderr, err := rollchainChain.Validators[i].ExecBin(ctx, "keys", "tiablob", "restore", "kick raven pave wild outdoor dismiss happy start lunch discover job evil code trim network emerge summer mad army vacant chest birth subject seek")
 		require.NoError(t, err, "stdout: %s, stderr: %s", stdout, stderr)
@@ -106,24 +106,24 @@ func watchForPublishedBlocks(
 	// Run test to observe x blocks posted
 	rollchainBlocksSeen := int64(0)
 	rollchainHighestBlock := int64(0)
-	celestiaHeight := uint64(1)
+	celestiaHeight := int64(1)
 
 	// setup time will allow publishedBlockCount to be used as a timeout
-	for i:=int64(0); i<publishedBlockCount; i++ {
+	for i := int64(0); i < publishedBlockCount; i++ {
 		err := testutil.WaitForBlocks(ctx, 1, rollchainChain)
 		require.NoError(t, err, "failed to wait for 1 block")
 
 		celestiaLatestHeight, err := celestiaChain.Height(ctx)
 		require.NoError(t, err, "error getting celestia height")
- 
+
 		for ; celestiaHeight < celestiaLatestHeight; celestiaHeight++ {
-			blobs, err := celestiaNodeClient.GetAllBlobs(ctx, celestiaHeight, "0x"+hex.EncodeToString([]byte("rc_demo")))
+			blobs, err := celestiaNodeClient.GetAllBlobs(ctx, uint64(celestiaHeight), "0x"+hex.EncodeToString([]byte("rc_demo")))
 			require.NoError(t, err, fmt.Sprintf("error getting all blobs at height: %d, %v", celestiaHeight, err))
 			t.Log("GetAllBlobs, celestia height: ", celestiaHeight)
 			if len(blobs) == 0 {
 				t.Log("No blobs found")
 			} else {
-				for j:=0; j<len(blobs); j++ {
+				for j := 0; j < len(blobs); j++ {
 					var block blocktypes.Block
 					err = block.Unmarshal(blobs[j].Data)
 					if err != nil {
